Use Gate::authenticate directly

+++ app/Http/Controllers/MailLogController.php
@@ -6,22 +6,17 @@ use App\Models\MailLog;
 use Illuminate\Contracts\View\View;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Gate;
 
 class MailLogController extends Controller
 {
-    /**
-     * Create the controller instance.
-     */
-    public function __construct()
-    {
-        $this->authorizeResource(MailLog::class, 'mail');
-    }
-
     /**
      * Display a listing of the resource.
      */
     public function index(): View
     {
+        Gate::authorize('viewAny', MailLog::class);
+
         return view('mail-log.index', [
             'mails' => MailLog::orderByDesc('created_at')->get(),
         ]);
@@ -32,6 +27,8 @@ class MailLogController extends Controller
      */
     public function show(MailLog $mail): View
     {
+        Gate::authorize('view', $mail);
+
         $mail->markRead()->save();
 
         return view('mail-log.show', [
@@ -41,7 +38,7 @@ class MailLogController extends Controller
 
     public function raw(MailLog $mail): Response
     {
-        $this->authorize('view', $mail);
+        Gate::authorize('view', $mail);
 
         return response($mail->body, headers: [
             'Content-Type' => 'text/plain; charset=utf-8',
@@ -54,6 +51,8 @@ class MailLogController extends Controller
      */
     public function destroy(MailLog $mail): RedirectResponse
     {
+        Gate::authorize('delete', $mail);
+
         $mail->delete();
 
         return redirect()->route('mail.index')
